Joint 2.9.2++
One extra fix for the system - changed the context text editor to prevent getting edited when users feed a text from a string table (external sources)

This is an extra change for the convenience. you don't need to update the system to this version at the moment if it's not crucial to you.